Deleted matrix_viewer_with_timing_fixed.py and mps_time_checking.py
ALSO merged changes into matrix_viewer_with_timing.py, modified README.md, added back mps_matrix_viewer_kathy.py, to be deleted later.
